requirements.txt
djangorestframework
astrapy
sentence-transformers
openai